Removed handle from mixnet client and introduced forwarder
